Updates directory structure to Go standard.
Moves source files into `src/eventsource-server` to allow for simply adding the
root of the checkedout source to the `GOPATH` environment variable for building
instead of requiering the folder structure on the local machine to be set up a
special way for Go.

Removes Makefile, the Go project page clearly state that any Makefile is not
needed for Go and that their inclusion is a failure of Go :)
https://golang.org/doc/articles/go_command.html second paragraph under
"Motivation". Also the Makefile was not used in the documented build anyway.

Updates the README to include new instructions for building the project.