Adds config files for Codeship CI/CD
Adds docker ignore file. Any path listed in it will not be available when adding
local files to a docker image using a Dockerfile.

Adds dockerfiles for building, includes go and git packages, distribution and
"deployment", as in uploading to our private Google cloud registry.

Adds encrypted credentials for Google cloud services. Details on how these are
built and used is in the Codeship docs. Basically it's some JSON config from
Google encrypted using a AES key from Codeship. When a new push happens Codeship
uses the AES key bound to the account to decrypt the credentials and then uses
them to connect to the Google cloud.

Adds the services configuration for Codeship. It's a list of Docker images,
existing or how to build them, that will be used in the build steps.

Adds the step definition file for Codeship. Each of these steps are run on the
build server. Any failing step blocks the following steps. If all steps pass
that means the image built correctly and got pushed to the cloud registry.