Replaces the CLI configuration with ENV variables
Replaces the `flag` package with the `os` package.

Replaces the use of `flag.` with corresponding `os.Getenv` to get the
configuration options from the environment. This is so that the configuration
can be easily set when running the service inside of a Docker container. Support
for ENV configuration is endemic with Docker and even if you *can* use
command line arguments, environment variables is a lot easier and cleaner.