Adds real implementation for connection metric
Implements the actual logic required to send metrics in Go. The implementation
is VERY verbose, considering it will just create a JSON object in the other end.
Compiler asses it atm. I'll implement the rest and then extract the common
parts.

WIP; I still need to implement the calls to ensure that a specific metric
actually exist before we try to push values to it. This only has to be done once
and I'm planning to put that into the initializer and keep it out of the
frequently called metrics functions.