Implements push for second metric + refactoring
Adds implementation to push the post event stats. It's almost the same as the
first one but with two data points per push.

Refactors the implementation and extracts a helper for creating the data points
(used three times) as well as one for sending the metrics once created (used two
times).

WIP; Still have to add the initializer to make sure the metrics exist at server
startup.